Allow fractional scale values for canvas files
The scale_x and scale_y fields were previously integers, but now they are floats. This allows
for more precise scaling of canvas files. The validation that scale values were between -1 and
1 has also been removed.

+++ src/api/canvas/models.py
@@ -14,8 +14,8 @@ class CanvasFile(ApiModel):
     width: float
     height: float
     z_index: int = Field(alias="zIndex")
-    scale_x: int = Field(alias="scaleX")
-    scale_y: int = Field(alias="scaleY")
+    scale_x: float = Field(alias="scaleX")
+    scale_y: float = Field(alias="scaleY")
 
     model_config: ClassVar[ConfigDict] = ConfigDict(
         populate_by_name=True, from_attributes=True
